+++ packages/ui/src/components/flex/index.jsx
@@ -1,4 +1,4 @@
-import { UIElement, classnames } from "@elf-framework/sapa";
+import { UIElement, classnames, useMemo } from "@elf-framework/sapa";
 
 import { registerComponent } from "../../utils/component";
 import { propertyMap } from "../../utils/propertyMap";
@@ -11,13 +11,19 @@ export class Flex extends UIElement {
       content,
       stack,
       wrap = false,
+      sameWidth = false,
     } = this.props;
 
+    const localClass = useMemo(() => {
+      return classnames("elf--flex", className, {
+        stack,
+        wrap,
+        "same-width": sameWidth,
+      });
+    }, [className, stack, wrap, sameWidth]);
+
     const styleObject = {
-      class: classnames("elf--flex", className, {
-        stack: stack,
-        wrap: wrap,
-      }),
+      class: localClass,
 
       style: {
         ...propertyMap(style, {}),