+++ apps/ui-dev/CHANGELOG.md
@@ -1,5 +1,11 @@
 # ui-dev
 
+## 1.0.34
+
+### Patch Changes
+
+- update layout documentation
+
 ## 1.0.33
 
 ### Patch Changes