+++ packages/design-system/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @elf-framework/design-system
 
+## 0.0.11
+
+### Patch Changes
+
+- update ui style
+
+- Updated dependencies []:
+  - @elf-framework/design-tokens@0.0.11
+
 ## 2.2.1
 
 ### Patch Changes