+++ packages/sapa/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @elf-framework/sapa
 
+## 0.0.10
+
+### Patch Changes
+
+- add flatTemplate function
+
+## 0.0.9
+
+### Patch Changes
+
+- modify fragment children to array
+
 ## 2.2.6
 
 ### Patch Changes